Fixing QuillFile::supportsEditing method
The calling of this method right after creating a QuillFile will return
true for gif file, which is wrong.
It happens because state is still normal, and it will only be readonly after
loading the filters. So, at the end of the method, when the filters are already loaded,
we can check again for the read only state.