Fixing HistoryXML to do not ignore filter option without value
Red-eye reduction filter sets a filter option without internal values,
(e.g. <FilterOption name="selection-area" type=9></FilterOption>) and it was
not being treated correctly by this part of the code.

This was causing supportsEditing() returns false for files that use
this filter (NB#213644), since the XML decode returns an error.

This patch should fix that.