Change string literals into static variables
Adds a new private header with const definitions for the strings that were
spread all over the code, so having repeated strings is avoided, and they
are easier to modify if needed. Although the saving is small, will also trim
some bytes from the data section of the library. Also, use constants for
filter names as defined in QuillImageFilter.