Pass --as-needed to linker to avoid uneeded libraries
By default the GNU linker as an entry in the DT_NEEDED table of ELF objects
for each of the libraries specified with -lfoobar in the command line, even
when no symbols from the library are used. The --as-needed does not add an
entry for unused libraries.

The final result is that the linker will not load (and link) unneeded
libraries at runtime, thus speeding up program startup.