Handle gracefully thumbnails that get deleted by external reasons.
If a thumbnail is removed by another process while it's still marked
as existing in the current process, a loading error gets emitted.
This can happen also when another process is editing and re-creating
the thumbnail just the same time than the loading occurs.

Instead of an laoding error, unregister the thumbnail and let it get
loaded again via normal loading (generating) procedures.